requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
learn==1.0.0
MarkupSafe==2.1.3
numpy==1.26.3
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
sckit==0.0.5
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
